requirements.txt
google-api-python-client
langchain
python-dotenv
langchain-google-genai
langchain-ollama